@@ -95,55 +95,66 @@ class SLAMonitor:
         cutoff_time = timezone.utcnow() - timedelta(hours=lookback_hours)
 
         # Total SLA misses
-        total_misses = session.query(SlaMiss).filter(
-            SlaMiss.timestamp >= cutoff_time,
-        ).count()
+        total_misses = (
+            session.query(SlaMiss)
+            .filter(
+                SlaMiss.timestamp >= cutoff_time,
+            )
+            .count()
+        )
 
         # SLA misses by DAG
-        misses_by_dag = session.query(
-            SlaMiss.dag_id,
-            func.count(SlaMiss.dag_id).label("miss_count")
-        ).filter(
-            SlaMiss.timestamp >= cutoff_time,
-        ).group_by(SlaMiss.dag_id).order_by(
-            func.count(SlaMiss.dag_id).desc()
-        ).limit(10).all()
+        misses_by_dag = (
+            session.query(SlaMiss.dag_id, func.count(SlaMiss.dag_id).label("miss_count"))
+            .filter(
+                SlaMiss.timestamp >= cutoff_time,
+            )
+            .group_by(SlaMiss.dag_id)
+            .order_by(func.count(SlaMiss.dag_id).desc())
+            .limit(10)
+            .all()
+        )
 
         # SLA misses by task
-        misses_by_task = session.query(
-            SlaMiss.dag_id,
-            SlaMiss.task_id,
-            func.count(SlaMiss.task_id).label("miss_count")
-        ).filter(
-            SlaMiss.timestamp >= cutoff_time,
-        ).group_by(SlaMiss.dag_id, SlaMiss.task_id).order_by(
-            func.count(SlaMiss.task_id).desc()
-        ).limit(10).all()
+        misses_by_task = (
+            session.query(SlaMiss.dag_id, SlaMiss.task_id, func.count(SlaMiss.task_id).label("miss_count"))
+            .filter(
+                SlaMiss.timestamp >= cutoff_time,
+            )
+            .group_by(SlaMiss.dag_id, SlaMiss.task_id)
+            .order_by(func.count(SlaMiss.task_id).desc())
+            .limit(10)
+            .all()
+        )
 
         # Notifications sent
-        notifications_sent = session.query(SlaMiss).filter(
-            SlaMiss.timestamp >= cutoff_time,
-            SlaMiss.notification_sent,
-        ).count()
+        notifications_sent = (
+            session.query(SlaMiss)
+            .filter(
+                SlaMiss.timestamp >= cutoff_time,
+                SlaMiss.notification_sent,
+            )
+            .count()
+        )
 
         # Emails sent
-        emails_sent = session.query(SlaMiss).filter(
-            SlaMiss.timestamp >= cutoff_time,
-            SlaMiss.email_sent,
-        ).count()
+        emails_sent = (
+            session.query(SlaMiss)
+            .filter(
+                SlaMiss.timestamp >= cutoff_time,
+                SlaMiss.email_sent,
+            )
+            .count()
+        )
 
         return {
             "period_hours": lookback_hours,
             "total_sla_misses": total_misses,
             "notifications_sent": notifications_sent,
             "emails_sent": emails_sent,
-            "top_dags_with_misses": [
-                {"dag_id": dag_id, "miss_count": count}
-                for dag_id, count in misses_by_dag
-            ],
+            "top_dags_with_misses": [{"dag_id": dag_id, "miss_count": count} for dag_id, count in misses_by_dag],
             "top_tasks_with_misses": [
-                {"dag_id": dag_id, "task_id": task_id, "miss_count": count}
-                for dag_id, task_id, count in misses_by_task
+                {"dag_id": dag_id, "task_id": task_id, "miss_count": count} for dag_id, task_id, count in misses_by_task
             ],
             "timestamp": timezone.utcnow().isoformat(),
         }
@@ -168,11 +179,15 @@ class SLAMonitor:
         cutoff_time = timezone.utcnow() - timedelta(days=lookback_days)
 
         # Get all completed runs for the DAG
-        dag_runs = session.query(DagRun).filter(
-            DagRun.dag_id == dag_id,
-            DagRun.end_date >= cutoff_time,
-            DagRun.end_date.isnot(None),
-        ).all()
+        dag_runs = (
+            session.query(DagRun)
+            .filter(
+                DagRun.dag_id == dag_id,
+                DagRun.end_date >= cutoff_time,
+                DagRun.end_date.isnot(None),
+            )
+            .all()
+        )
 
         if not dag_runs:
             return {
@@ -207,10 +222,14 @@ class SLAMonitor:
         p95_idx = int(len(sorted_durations) * 0.95)
 
         # SLA misses for this DAG
-        sla_miss_count = session.query(SlaMiss).filter(
-            SlaMiss.dag_id == dag_id,
-            SlaMiss.timestamp >= cutoff_time,
-        ).count()
+        sla_miss_count = (
+            session.query(SlaMiss)
+            .filter(
+                SlaMiss.dag_id == dag_id,
+                SlaMiss.timestamp >= cutoff_time,
+            )
+            .count()
+        )
 
         return {
             "dag_id": dag_id,
@@ -246,9 +265,13 @@ class SLAMonitor:
         from airflow.utils.state import DagRunState
 
         # Get currently running DAG runs
-        running_dags = session.query(DagRun).filter(
-            DagRun.state == DagRunState.RUNNING,
-        ).all()
+        running_dags = (
+            session.query(DagRun)
+            .filter(
+                DagRun.state == DagRunState.RUNNING,
+            )
+            .all()
+        )
 
         at_risk = []
         current_time = timezone.utcnow()
@@ -259,14 +282,16 @@ class SLAMonitor:
 
                 # Check if running longer than threshold
                 if running_time > sla_threshold_minutes:
-                    at_risk.append({
-                        "dag_id": dag_run.dag_id,
-                        "run_id": dag_run.run_id,
-                        "execution_date": dag_run.execution_date.isoformat(),
-                        "start_date": dag_run.start_date.isoformat(),
-                        "running_minutes": round(running_time, 2),
-                        "threshold_minutes": sla_threshold_minutes,
-                        "risk_level": "high" if running_time > sla_threshold_minutes * 2 else "medium",
-                    })
+                    at_risk.append(
+                        {
+                            "dag_id": dag_run.dag_id,
+                            "run_id": dag_run.run_id,
+                            "execution_date": dag_run.execution_date.isoformat(),
+                            "start_date": dag_run.start_date.isoformat(),
+                            "running_minutes": round(running_time, 2),
+                            "threshold_minutes": sla_threshold_minutes,
+                            "risk_level": "high" if running_time > sla_threshold_minutes * 2 else "medium",
+                        }
+                    )
 
         return at_risk
