@@ -51,16 +51,18 @@ class TaskHealthMonitor:
         for ti in query.all():
             running_minutes = (current_time - ti.start_date).total_seconds() / 60
 
-            long_running.append({
-                "dag_id": ti.dag_id,
-                "task_id": ti.task_id,
-                "run_id": ti.run_id,
-                "execution_date": ti.execution_date.isoformat() if ti.execution_date else None,
-                "start_date": ti.start_date.isoformat() if ti.start_date else None,
-                "running_minutes": round(running_minutes, 2),
-                "operator": ti.operator,
-                "severity": self._get_duration_severity(running_minutes, threshold_minutes),
-            })
+            long_running.append(
+                {
+                    "dag_id": ti.dag_id,
+                    "task_id": ti.task_id,
+                    "run_id": ti.run_id,
+                    "execution_date": ti.execution_date.isoformat() if ti.execution_date else None,
+                    "start_date": ti.start_date.isoformat() if ti.start_date else None,
+                    "running_minutes": round(running_minutes, 2),
+                    "operator": ti.operator,
+                    "severity": self._get_duration_severity(running_minutes, threshold_minutes),
+                }
+            )
 
         return sorted(long_running, key=lambda x: x["running_minutes"], reverse=True)
 
@@ -95,23 +97,29 @@ class TaskHealthMonitor:
         cutoff_time = timezone.utcnow() - timedelta(hours=lookback_hours)
 
         # Get tasks that have been retried multiple times
-        query = session.query(TaskInstance).filter(
-            TaskInstance.end_date >= cutoff_time,
-            TaskInstance.try_number > min_retries,
-        ).order_by(TaskInstance.try_number.desc())
+        query = (
+            session.query(TaskInstance)
+            .filter(
+                TaskInstance.end_date >= cutoff_time,
+                TaskInstance.try_number > min_retries,
+            )
+            .order_by(TaskInstance.try_number.desc())
+        )
 
         retry_tasks = []
         for ti in query.limit(50).all():
-            retry_tasks.append({
-                "dag_id": ti.dag_id,
-                "task_id": ti.task_id,
-                "run_id": ti.run_id,
-                "try_number": ti.try_number,
-                "max_tries": ti.max_tries,
-                "state": str(ti.state),
-                "operator": ti.operator,
-                "last_attempt": ti.end_date.isoformat() if ti.end_date else None,
-            })
+            retry_tasks.append(
+                {
+                    "dag_id": ti.dag_id,
+                    "task_id": ti.task_id,
+                    "run_id": ti.run_id,
+                    "try_number": ti.try_number,
+                    "max_tries": ti.max_tries,
+                    "state": str(ti.state),
+                    "operator": ti.operator,
+                    "last_attempt": ti.end_date.isoformat() if ti.end_date else None,
+                }
+            )
 
         return retry_tasks
 
@@ -143,26 +151,32 @@ class TaskHealthMonitor:
 
         for ti in query.all():
             # Check if the task's DAG run is still running
-            dag_run = session.query(DagRun).filter(
-                DagRun.dag_id == ti.dag_id,
-                DagRun.run_id == ti.run_id,
-            ).first()
+            dag_run = (
+                session.query(DagRun)
+                .filter(
+                    DagRun.dag_id == ti.dag_id,
+                    DagRun.run_id == ti.run_id,
+                )
+                .first()
+            )
 
             # If DAG run is not running but task is, it's likely a zombie
             is_orphaned = dag_run and dag_run.state != DagRunState.RUNNING
 
             running_minutes = (current_time - ti.start_date).total_seconds() / 60
 
-            zombies.append({
-                "dag_id": ti.dag_id,
-                "task_id": ti.task_id,
-                "run_id": ti.run_id,
-                "start_date": ti.start_date.isoformat() if ti.start_date else None,
-                "running_minutes": round(running_minutes, 2),
-                "is_orphaned": is_orphaned,
-                "dag_run_state": str(dag_run.state) if dag_run else "unknown",
-                "operator": ti.operator,
-            })
+            zombies.append(
+                {
+                    "dag_id": ti.dag_id,
+                    "task_id": ti.task_id,
+                    "run_id": ti.run_id,
+                    "start_date": ti.start_date.isoformat() if ti.start_date else None,
+                    "running_minutes": round(running_minutes, 2),
+                    "is_orphaned": is_orphaned,
+                    "dag_run_state": str(dag_run.state) if dag_run else "unknown",
+                    "operator": ti.operator,
+                }
+            )
 
         return zombies
 
@@ -184,59 +198,75 @@ class TaskHealthMonitor:
         cutoff_time = timezone.utcnow() - timedelta(hours=lookback_hours)
 
         # Get failure counts by task
-        failure_counts = session.query(
-            TaskInstance.dag_id,
-            TaskInstance.task_id,
-            func.count(TaskInstance.task_id).label("failure_count"),
-        ).filter(
-            TaskInstance.state == TaskInstanceState.FAILED,
-            TaskInstance.end_date >= cutoff_time,
-        ).group_by(
-            TaskInstance.dag_id,
-            TaskInstance.task_id,
-        ).order_by(
-            func.count(TaskInstance.task_id).desc()
-        ).limit(20).all()
+        failure_counts = (
+            session.query(
+                TaskInstance.dag_id,
+                TaskInstance.task_id,
+                func.count(TaskInstance.task_id).label("failure_count"),
+            )
+            .filter(
+                TaskInstance.state == TaskInstanceState.FAILED,
+                TaskInstance.end_date >= cutoff_time,
+            )
+            .group_by(
+                TaskInstance.dag_id,
+                TaskInstance.task_id,
+            )
+            .order_by(func.count(TaskInstance.task_id).desc())
+            .limit(20)
+            .all()
+        )
 
         # Get total task runs for failure rate calculation
-        total_runs = session.query(TaskInstance).filter(
-            TaskInstance.end_date >= cutoff_time,
-        ).count()
+        total_runs = (
+            session.query(TaskInstance)
+            .filter(
+                TaskInstance.end_date >= cutoff_time,
+            )
+            .count()
+        )
 
-        total_failures = session.query(TaskInstance).filter(
-            TaskInstance.state == TaskInstanceState.FAILED,
-            TaskInstance.end_date >= cutoff_time,
-        ).count()
+        total_failures = (
+            session.query(TaskInstance)
+            .filter(
+                TaskInstance.state == TaskInstanceState.FAILED,
+                TaskInstance.end_date >= cutoff_time,
+            )
+            .count()
+        )
 
         # Identify flaky tasks (high failure + high retry success)
         flaky_tasks = []
         for dag_id, task_id, failure_count in failure_counts[:10]:
-            success_after_retry = session.query(TaskInstance).filter(
-                TaskInstance.dag_id == dag_id,
-                TaskInstance.task_id == task_id,
-                TaskInstance.state == TaskInstanceState.SUCCESS,
-                TaskInstance.try_number > 1,
-                TaskInstance.end_date >= cutoff_time,
-            ).count()
+            success_after_retry = (
+                session.query(TaskInstance)
+                .filter(
+                    TaskInstance.dag_id == dag_id,
+                    TaskInstance.task_id == task_id,
+                    TaskInstance.state == TaskInstanceState.SUCCESS,
+                    TaskInstance.try_number > 1,
+                    TaskInstance.end_date >= cutoff_time,
+                )
+                .count()
+            )
 
             if success_after_retry > 0:
-                flaky_tasks.append({
-                    "dag_id": dag_id,
-                    "task_id": task_id,
-                    "failure_count": failure_count,
-                    "retry_success_count": success_after_retry,
-                    "flakiness_score": round(success_after_retry / (failure_count + success_after_retry) * 100, 2),
-                })
+                flaky_tasks.append(
+                    {
+                        "dag_id": dag_id,
+                        "task_id": task_id,
+                        "failure_count": failure_count,
+                        "retry_success_count": success_after_retry,
+                        "flakiness_score": round(success_after_retry / (failure_count + success_after_retry) * 100, 2),
+                    }
+                )
 
         return {
             "period_hours": lookback_hours,
             "total_task_runs": total_runs,
             "total_failures": total_failures,
             "failure_rate": round(total_failures / total_runs * 100, 2) if total_runs > 0 else 0,
-            "top_failing_tasks": [
-                {"dag_id": d, "task_id": t, "failure_count": c}
-                for d, t, c in failure_counts
-            ],
+            "top_failing_tasks": [{"dag_id": d, "task_id": t, "failure_count": c} for d, t, c in failure_counts],
             "flaky_tasks": flaky_tasks,
             "timestamp": timezone.utcnow().isoformat(),
         }
