@@ -16,6 +16,7 @@ sys.path.insert(0, str(src_path))
 # install in the test environment.
 # ---------------------------------------------------------------------------
 
+
 def _make_stub(name):
     stub = MagicMock()
     stub.__name__ = name
@@ -30,13 +31,16 @@ _airflow_conf.getboolean = lambda s, k, fallback=None: fallback
 _timezone = MagicMock()
 _timezone.utcnow = datetime.utcnow
 
+
 # provide_session: just call the function directly with session=None
 def _provide_session(f):
     def wrapper(*args, **kwargs):
         kwargs.setdefault("session", MagicMock())
         return f(*args, **kwargs)
+
     wrapper.__name__ = getattr(f, "__name__", "wrapped")
     return wrapper
+
 
 _STUBS = {
     "airflow": MagicMock(),
@@ -72,20 +76,36 @@ class _ColMock(MagicMock):
     calls (e.g. DagRun.end_date >= cutoff). Without this, MagicMock raises
     TypeError when compared to a real datetime.
     """
-    def __ge__(self, other): return MagicMock()
-    def __gt__(self, other): return MagicMock()
-    def __le__(self, other): return MagicMock()
-    def __lt__(self, other): return MagicMock()
-    def __eq__(self, other): return MagicMock()
-    def __ne__(self, other): return MagicMock()
+
+    def __ge__(self, other):
+        return MagicMock()
+
+    def __gt__(self, other):
+        return MagicMock()
+
+    def __le__(self, other):
+        return MagicMock()
+
+    def __lt__(self, other):
+        return MagicMock()
+
+    def __eq__(self, other):
+        return MagicMock()
+
+    def __ne__(self, other):
+        return MagicMock()
 
 
 def _make_model_mock():
     """Create a MagicMock where attribute access returns _ColMock instances."""
     m = MagicMock()
-    m.__class__ = type("ModelMock", (MagicMock,), {
-        "__getattr__": lambda self, name: _ColMock(),
-    })
+    m.__class__ = type(
+        "ModelMock",
+        (MagicMock,),
+        {
+            "__getattr__": lambda self, name: _ColMock(),
+        },
+    )
     return m
 
 
@@ -98,6 +118,7 @@ for _model_name in ["DagRun", "TaskInstance", "DagModel", "SlaMiss", "Pool"]:
 # ---------------------------------------------------------------------------
 # Shared fixtures
 # ---------------------------------------------------------------------------
+
 
 @pytest.fixture
 def sample_datetime():
